+++ app/classes/controllers/management_controller.py
@@ -95,9 +95,6 @@ class ManagementController:
     # **********************************************************************************
     #                                   Audit_Log Methods
     # **********************************************************************************
-    @staticmethod
-    def get_activity_log():
-        return HelpersManagement.get_activity_log()
 
     def add_to_audit_log(self, user_id, log_msg, server_id=None, source_ip=None):
         return self.management_helper.add_to_audit_log(