+++ app/classes/web/base_handler.py
@@ -182,6 +182,7 @@ class BaseHandler(tornado.web.RequestHandler):
             t.List[str],
             bool,
             t.Dict[str, t.Any],
+            str,
         ]
     ]:
         try:
@@ -190,9 +191,10 @@ class BaseHandler(tornado.web.RequestHandler):
             )
 
             superuser = user["superuser"]
+            server_permissions_api_mask = ""
             if api_key is not None:
-                superuser = superuser and api_key.superuser
-
+                superuser = superuser and api_key.full_access
+                server_permissions_api_mask = api_key.server_permissions
             exec_user_role = set()
             if superuser:
                 authorized_servers = self.controller.servers.get_all_defined_servers()
@@ -214,6 +216,7 @@ class BaseHandler(tornado.web.RequestHandler):
                             user["user_id"]
                         )
                     )
+
                 logger.debug(user["roles"])
                 for r in user["roles"]:
                     role = self.controller.roles.get_role(r)
@@ -234,6 +237,7 @@ class BaseHandler(tornado.web.RequestHandler):
                     exec_user_role,
                     superuser,
                     user,
+                    server_permissions_api_mask,
                 )
             logging.debug("Auth unsuccessful")
             auth_log.error(