+++ app/classes/web/routes/api/servers/index.py
@@ -139,7 +139,7 @@ new_server_schema = {
                     "category": {
                         "title": "Jar Category",
                         "type": "string",
-                        "examples": ["modded", "vanilla"],
+                        "examples": ["Mc_java_servers", "Mc_java_proxies"],
                     },
                     "properties": {
                         "type": {
@@ -685,6 +685,7 @@ class ApiServersIndexHandler(BaseApiHandler):
             _,
             _superuser,
             user,
+            _,
         ) = auth_data
 
         if EnumPermissionsCrafty.SERVER_CREATION not in exec_user_crafty_permissions: